+++ src/main/kotlin/com/deepromeet/seulseul/user/api/request/UserInfoUpdateRequest.kt
@@ -1,0 +1,9 @@
+package com.deepromeet.seulseul.user.api.request
+
+data class UserInfoUpdateRequest(
+    val nickname: String,
+    val thumbnailImageUrl: String,
+    val profileImageUrl: String,
+    val alertAgreement: Boolean,
+    val trackingAgreement: Boolean,
+)
